Revert to using the simple-water component but refactor the Simple Water Mesh
