+++ src/aframe-to-bit-components.js
@@ -8,8 +8,7 @@ import {
   NotRemoteHoverTarget,
   RemoveNetworkedEntityButton,
   DestroyAtExtremeDistance,
-  Billboard,
-  SimpleWater
+  Billboard
 } from "./bit-components";
 
 [
@@ -21,8 +20,7 @@ import {
   ["is-not-remote-hover-target", NotRemoteHoverTarget],
   ["remove-networked-object-button", RemoveNetworkedEntityButton],
   ["destroy-at-extreme-distances", DestroyAtExtremeDistance],
-  ["billboard", Billboard],
-  ["simple-water", SimpleWater]
+  ["billboard", Billboard]
 ].forEach(([aframeComponentName, bitecsComponent]) => {
   AFRAME.registerComponent(aframeComponentName, {
     init: function () {
